refactor: 오타 수정 & 모달 client
