+++ src/features/(public)/sign/useGoogleSignin.ts
@@ -40,7 +40,6 @@ export function useGoogleSignin() {
         const sp = new URLSearchParams({
           title: '승인 대기중 입니다.',
           message: '관리자가 승인을 완료하면 서비스를 이용가능합니다.',
-          returnTo: '/signin',
         })
 
         router.push(`/signin/message?${sp.toString()}`)
